refactor(settings): extract account sections and give agent stats its own card
Split the account settings page into dedicated components under
components/settings/ (profile, login-details, connected-accounts,
privacy, chat) so page.tsx is just composition. Move the agent stats
toggle out of Privacy into its own Chat section and align the loading
skeleton with the real layout (PageContainer max-width, toggle/provider
row placeholders, all seven cards).
